Migration: Field type registry, version history, and portal fixes
- Add field type registry API endpoints (/field-types, /field-types/toolbox)
- Integrate FieldNormalizer and VersionService into form submissions
- UpdateFormStructure now populates field_type_id from registry
- UpdateTableRow creates version history with field change tracking
- FieldToolbox fetches field types from API with static fallback
- Fix form submission persistence (is_published flag)
- Fix PublicPortal port fallback (8000 -> 8080)
- Fix login redirect to verify workspace membership
- Add autocomplete attributes to login form inputs
- Clean up debug logs from handlers

+++ go-backend/handlers/data_tables.go
@@ -215,7 +215,7 @@ func ListTableRows(c *gin.Context) {
 
 type CreateTableRowInput struct {
 	Data     map[string]interface{} `json:"data" binding:"required"`
-	Position int                    `json:"position"`
+	Position int64                  `json:"position"`
 }
 
 func CreateTableRow(c *gin.Context) {
@@ -294,7 +294,7 @@ func CreateTableRow(c *gin.Context) {
 
 type UpdateTableRowInput struct {
 	Data     *map[string]interface{} `json:"data"`
-	Position *int                    `json:"position"`
+	Position *int64                  `json:"position"`
 }
 
 func UpdateTableRow(c *gin.Context) {
